+++ ChangeLog.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 
 ## 0.1.3 - ?
 
+### Added
+* New `DownloadBegin` event.
+
 ### Fixed
 * Race condition for message encoding.
   The encoding buffer is now external to `codec::Encoder`.
@@ -13,6 +16,7 @@ All notable changes to this project will be documented in this file.
 * Changed `Ready` to include `max_order_id` (highest order id previously used by the client).
 * Changed `MarketByPrice` and `TradeSummary` field `channel` to `uint16_t`.
 * Renamed `MessageInfo` field `is_cache` to `from_cache`.
+* Renamed `Ready` to `DownloadEnd`
 
 
 ## 0.1.2 - 2018-03-31